requirements.txt
# Data Analysis and Visualization
pandas>=1.3.5
numpy>=1.21.5
matplotlib>=3.4.3
seaborn>=0.11.2
streamlit>=1.33.0
plotly-express==0.4.1


# Unit Testing
pytest>=7.2.0
coverage>=6.3.2

# Continuous Integration (CI)
pytest-cov>=3.0.0
codecov==2.1.13

# Additional Utilities
scikit-learn>=1.0.2